refactor code + update sonar properties
