--- dm_backend/models/__init__.py
@@ -1,5 +1,0 @@
-"""This is the models module of the dm_backend package."""
-from dm_backend.models.equation import Equation
-from dm_backend.models.graph import Graph
-
-__all__ = ["Equation", "Graph"]
