+++ dm_backend/src/models/__init__.py
@@ -1,0 +1,5 @@
+"""This is the models module of the dm_backend package."""
+from dm_backend.src.models.equation import Equation
+from dm_backend.src.models.graph import Graph
+
+__all__ = ["Equation", "Graph"]
