+++ dm_backend/tests/baker_recipes/equation_baker_recipe.py
@@ -1,5 +1,5 @@
 from model_bakery.recipe import Recipe
 
-from dm_backend.models import Equation
+from dm_backend.src.models import Equation
 
 equation_recipe = Recipe(Equation)
