+++ dm_backend/tests/baker_recipes/graph_baker_recipe.py
@@ -1,5 +1,5 @@
 from model_bakery.recipe import Recipe
 
-from dm_backend.models import Graph
+from dm_backend.src.models import Graph
 
 graph_recipe = Recipe(Graph)
