+++ sonar-project.properties
@@ -10,7 +10,6 @@ sonar.organization=swe-muic
 
 # Encoding of the source code. Default is default system encoding
 #sonar.sourceEncoding=UTF-8
-sonar.sources=dm_backend
-; sonar.tests=dm_backend.tests
-; sonar.exclustions=dm_backend.tests/**/*
+sonar.sources=dm_backend.src
+sonar.tests=dm_backend.tests
 sonar.python.coverage.reportPaths=coverage.xml
